pyproject.toml
[project]
name = "dockerscope"
version = "0.1.0"
description = "Analyze Docker environments for real security risks and attack paths."
readme = "README.md"
requires-python = ">=3.11"

authors = [
  { name = "Tal Hayun" }
]

dependencies = [
  "docker>=7.0.0",
  "typer[all]>=0.12.0",
  "networkx>=3.0",
  "PyYAML>=6.0",
  "rich>=13.0.0",
]

[project.scripts]
dockerscope = "dockerscope.cli:main"

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=5.0.0",
  "pytest-mock>=3.14.0",
  "ruff>=0.4.0",
  "mypy>=1.10.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
